Updated
Changed to numpy arrays instead of lists. Found that it was easier to index
Added a dipole2 function that can take into account n unit cells. Slight issue with it including bars that aren't quite in the unit cell but they do make a sizable contribution to the local field. This new function can also handle edges
Added in the kagome lattice builder. Haven't tested it with dipole interaction yet.

Can meet up today if you like to discuss this