Hysteresis
Added Hysteresis to the field sweep so that is returns the net magnetisation of the array after each field step.
The net magnetisation takes the average magnetisation in the x and y direction and returns it as an two numbers.
Added an example graph of Hysteresis for a square array
I also commented out some of the graphs